清理代码
