+++ Assets/Source/Common/Tools/MouseDetectionTool.cs.meta
