+++ Assets/Source/Common/Tools/Timer.cs
