+++ Assets/Source/Common/UI/UIViewBase.cs
