+++ Assets/Source/Controller/MainFSMCommand.cs
@@ -15,27 +15,7 @@ public class MainFSMCommand : SimpleCommand
 
         switch (name)
         {
-            case Const.Notification.LOGIN_SUCCESS:
-                //SceneManager.LoadScene(Const.SceneNames.MAIN_PANEL_SCENE);
-                //SendNotification(Const.Notification.REQUEST_FOR_GAME_STATUS);
-                break;
-                /*
-            case Const.Notification.RECEIVED_GAME_STATUS:
-                if ((obj as GameStatusVO).gameStatus == GameStatus.s)
-                {
-                    GameManager.instance.ChangeMainFSMState(MainFSMStateID.InGame);
-                }
-                else if ((obj as GameStatusVO).gameStatus == GameStatus.p)
-                {
-                    GameManager.instance.ChangeMainFSMState(MainFSMStateID.PreGame);
-                }
-                //TODO: Test only, need to remove later
-                else
-                {
-                    GameManager.instance.ChangeMainFSMState(MainFSMStateID.PreGame);
-                }
-                break;
-                */               
+                         
         }
     }
 }
