--- Assets/Source/Controller/MapCommand.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2689d90df5fe3464db62af63d0df9535
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
