+++ Assets/Source/Model/ChangeGameStatusProxy.cs
@@ -30,7 +30,6 @@ public class ChangeGameStatusProxy : Proxy, IProxy, IResponder
         {
             SendNotification(Const.Notification.GAME_STARTED);
         }
-
         if ((_data as ChangeGameStatusResponse).device_id_2_user_info != null)
         {
             SendNotification(Const.Notification.UPDATE_DEVICE_ID_TO_USER_INFO, (_data as ChangeGameStatusResponse).device_id_2_user_info);
