+++ Assets/Source/View/DebugViewMediator.cs
@@ -102,7 +102,6 @@ public class DebugViewMediator : Mediator, IMediator
         {
 
         }
-
     }
     
 }
