+++ Assets/Source/View/GameStatusViewMediator.cs
@@ -48,7 +48,6 @@ public class GameStatusViewMediator : Mediator, IMediator
                 UpdateGamesessioInfo(vo as GameSessionsResponse);
                 break;
             case Const.Notification.UPDATE_DEVICE_ID_TO_USER_INFO:
-                SendNotification(Const.Notification.DEBUG_LOG, "Player Updated: " + (vo as Dictionary<string, UserInfo>).Count);
 
                 //TODO: Remove foreach{} when finished testing
                 foreach(KeyValuePair<string, UserInfo> kvp in (vo as Dictionary<string, UserInfo>))
@@ -56,7 +55,8 @@ public class GameStatusViewMediator : Mediator, IMediator
                     PlayerPosSImulator.instance.OnNewVirtualPlayer(kvp.Key);
                 }
 
-                AppFacade.instance.SendNotification(Const.Notification.DEBUG_LOG, "Players updated: " + (vo as Dictionary<string, UserInfo>).Count.ToString());
+
+                SendNotification(Const.Notification.DEBUG_LOG, "Player Updated: " + (vo as Dictionary<string, UserInfo>).Count);
                 break;
         }
     }
