--- Assets/Source/View/MainView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 91b3ba884d3523a4b9fa49e9c8c4a804
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
