--- Assets/Source/View/MainViewMediator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 59cfb1c799c3068418cee5108262041d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
