+++ Assets/Source/View/McuItemViewMediator.cs
@@ -32,7 +32,6 @@ public class McuItemViewMediator : Mediator, IMediator
 
         switch (name)
         {
-
         }
     }
